Hotfix float

+++ app/controllers/recipes.py
@@ -183,7 +183,7 @@ class RecipesView(BaseRecipesView, ExtendedFlaskView):
         self.ingredient = Ingredient.load(ingredient_id)
         amount = request.form["amount"]
 
-        amount_for_portion = int(amount) / float(self.recipe.portion_count)
+        amount_for_portion = float(amount) / float(self.recipe.portion_count)
 
         self.recipe.change_ingredient_amount(self.ingredient, amount_for_portion)
         self.ingredient.amount = amount_for_portion
